refactor: merge descriptor into type.py
